+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [1.0.14](https://github.com/jorigin/jcommon/releases/tag/release-1.0.14)
 ### Changed
- - IconLoadr can now load icons from Jar files or specified directory
+ - Refactored package `org.jorigin.gui` into `org.jorigin.swing`
+ - Moved class GraphicsEnvironmentPanel from `org.jorigin.system.gui` to `org.jorigin.swing`
+ - IconLoader can now load icons from Jar files or specified directory
 
 ## [1.0.13](https://github.com/jorigin/jcommon/releases/tag/release-1.0.13)
 ### Changed